requirements.txt
langchain
flask==3.1.1 
sentence-transformers==4.1.0
pypdf==5.6.1 
python-dotenv==1.1.0
langchain-pinecone
google-generativeai
langchain-community
langchain-google-genai
langchain-groq

groq
-e .